calls: Update return type in lookup_dcookie()

+++ src/platform/linux-s390x/call.rs
@@ -3991,12 +3991,11 @@ pub unsafe fn llistxattr<P: AsRef<Path>>(filename: P, value: &mut [u8]) -> Resul
 }
 
 /// Return a directory entry's path.
-// TODO(Shaohua): Returns a string.
-pub unsafe fn lookup_dcookie(cookie: u64, buf: &mut [u8]) -> Result<i32, Errno> {
+pub unsafe fn lookup_dcookie(cookie: u64, buf: &mut [u8]) -> Result<ssize_t, Errno> {
     let cookie = cookie as usize;
     let buf_ptr = buf.as_mut_ptr() as usize;
     let buf_len = buf.len();
-    syscall3(SYS_LOOKUP_DCOOKIE, cookie, buf_ptr, buf_len).map(|ret| ret as i32)
+    syscall3(SYS_LOOKUP_DCOOKIE, cookie, buf_ptr, buf_len).map(|ret| ret as ssize_t)
 }
 
 /// Remove an extended attribute.
